requirements.txt
streamlit>=1.22.0
sentence-transformers>=2.2.2
numpy>=1.22.0
scikit-learn>=1.1.0
torch>=1.12.0
requests>=2.28.1
python-dotenv>=0.19.0
nltk>=3.8.0
faiss-cpu>=1.7.3
flask>=2.0.0
flask-cors>=3.0.0
PyPDF2>=3.0.0
beautifulsoup4>=4.11.0
